fix: MSVC/Windows cross-platform compatibility for v1/v2/v3
v1 fixes:
- CMakeLists.txt: use if(MSVC) for compile/link flags, add /utf-8
- MemoryPool.h: change BlockSize_/SlotSize_/getMemoryPool to size_t
- UnitTest.cpp: use %zu for size_t in printf

v2/v3 fixes:
- CMakeLists.txt: platform-conditional compile options, fix CMAKE_SOURCE_DIR paths, fix Unix-style test commands
- PageCache.cpp: replace sys/mman.h + mmap with VirtualAlloc on Windows
- UnitTest.cpp/PerformanceTest.cpp: fix MSVC lambda capture requirements

+++ v1/include/MemoryPool.h
@@ -39,8 +39,8 @@ private:
     bool pushFreeList(Slot* slot);
     Slot* popFreeList();
 private:
-    int                 BlockSize_; // 内存块大小
-    int                 SlotSize_; // 槽大小
+    size_t              BlockSize_; // 内存块大小
+    size_t              SlotSize_; // 槽大小
     Slot*               firstBlock_; // 指向内存池管理的首个实际内存块
     Slot*               curSlot_; // 指向当前未被使用过的槽
     std::atomic<Slot*>  freeList_; // 指向空闲的槽(被使用过后又被释放的槽)
@@ -53,7 +53,7 @@ class HashBucket
 {
 public:
     static void initMemoryPool();
-    static MemoryPool& getMemoryPool(int index);
+    static MemoryPool& getMemoryPool(size_t index);
 
     static void* useMemory(size_t size)
     {
